chore: rename theme context components
