--- src/contexts/chop-logic-form-context.ts
@@ -1,4 +1,0 @@
-import React from 'react';
-import { ChopLogicFormContextProps } from '@models';
-
-export const ChopLogicFormContext = React.createContext<ChopLogicFormContextProps>({});
