+++ src/contexts/index.ts
@@ -1,1 +1,3 @@
-export { ChopLogicFormContext } from './chop-logic-form-context';
+export { ChopLogicFormContext } from './form/context.tsx';
+export { ChopLogicThemeContext } from './theme/context.ts';
+export { ChopLogicThemeProvider } from './theme/provider.tsx';
