+++ src/models/index.ts
@@ -1,4 +1,3 @@
-export type { ChopLogicTheme, ChopLogicThemeContextProps } from './chop-logic-theme.ts';
 export type {
   ChopLogicComponentProps,
   ChopLogicInputProps,
@@ -12,13 +11,7 @@ export type { ChopLogicAlertProps } from './components/alert.ts';
 export type { ChopLogicButtonProps } from './components/button.ts';
 export type { ChopLogicCheckboxProps } from './components/checkbox.ts';
 export type { ChopLogicDialogProps } from './components/dialog.ts';
-export type {
-  ChopLogicFormContextProps,
-  ChopLogicFormData,
-  ChopLogicFormInputParams,
-  ChopLogicFormProps,
-  ChopLogicFormValidationState,
-} from './components/form.ts';
+export type { ChopLogicFormData, ChopLogicFormInputParams, ChopLogicFormProps, ChopLogicFormValidationState } from './components/form.ts';
 export type { ChopLogicGridColumn, ChopLogicGridItem, ChopLogicGridProps, RenderDataItemCallback } from './components/grid.ts';
 export type { ChopLogicIconProps, ChopLogicSVGElementProps } from './components/icon.ts';
 export type { ChopLogicLoaderProps } from './components/loader.ts';
