+++ src/styles/storybook/Container.module.scss
@@ -5,7 +5,7 @@
   color: var(--cl-font-color-base);
 
   padding: var(--cl-gap-big);
-  cursor: pointer;
+  margin: var(--cl-gap-big);
   border: var(--cl-border-light);
   display: flex;
   flex-direction: column;
